requirements.txt
matplotlib
torch
torchvision
numpy
pillow
pycocotools